requirements.txt
openai==1.59.3
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==1.0.0